# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Open VoIP API
  version: 1.1.0
  description: |-
    内网呼叫中心 REST API。WebSocket 见 ../events.md。
    需求 ID 见各 operation 的 x-requirement-id。

    JSON 响应统一为 code/message/data/request_id，详见 response-contract.md。原 204 操作现在返回 200 和 data:null；下载、存活探针和 WebSocket 升级除外。
    所有时间戳使用 UTC 的 YYYY-MM-DD HH:MM:SS，例如 2026-09-25 10:12:35。
  license:
    name: Proprietary
servers:
  - url: https://cc.internal
    description: 内网示例
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/calls/outbound:
    post:
      tags: [ Calls ]
      summary: 外呼
      x-requirement-id: CALL-01,MEDIA-09
      x-roles: [ agent ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutboundCallRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Call"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/answer:
    post:
      tags: [ Calls ]
      x-requirement-id: QUEUE-03
      x-roles: [ agent ]
      parameters:
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Call"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/hangup:
    post:
      tags: [ Calls ]
      x-roles: [ agent, supervisor ]
      parameters:
        - $ref: "#/components/parameters/CallId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/hold:
    post:
      tags: [ Calls ]
      x-requirement-id: MEDIA-05
      x-roles: [ agent ]
      parameters:
        - $ref: "#/components/parameters/CallId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                on:
                  type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/transfer:
    post:
      tags: [ Calls ]
      x-requirement-id: CALL-02,VIDEO-09
      x-roles: [ agent ]
      parameters:
        - $ref: "#/components/parameters/CallId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
        "422":
          $ref: "#/components/responses/BusinessError"
  /api/v1/calls/{callId}/transfer/complete:
    post:
      tags: [ Calls ]
      summary: 完成咨询转
      x-requirement-id: CALL-02
      x-roles: [ agent ]
      parameters:
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
        "409":
          $ref: "#/components/responses/BusinessError"
  /api/v1/calls/{callId}/video/request:
    post:
      tags: [ Calls ]
      x-requirement-id: VIDEO-06
      x-roles: [ agent ]
      parameters:
        - $ref: "#/components/parameters/CallId"
      responses:
        "202":
          description: 已通知对端
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/calls/{callId}:
    get:
      tags: [ Calls ]
      parameters:
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Call"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/decline:
    post:
      tags: [ Calls ]
      x-roles: [ agent ]
      parameters:
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/video/respond:
    post:
      tags: [ Calls ]
      x-requirement-id: VIDEO-06
      parameters:
        - $ref: "#/components/parameters/CallId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accept:
                  type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/video/downgrade:
    post:
      tags: [ Calls ]
      x-requirement-id: VIDEO-07
      parameters:
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/screen-share:
    post:
      tags: [ Calls ]
      x-requirement-id: VIDEO-08
      parameters:
        - $ref: "#/components/parameters/CallId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                on:
                  type: boolean
                leg_id:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/conference:
    post:
      tags: [ Calls ]
      x-requirement-id: CALL-03
      x-roles: [ agent ]
      parameters:
        - $ref: "#/components/parameters/CallId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [ agent_id ]
              properties:
                agent_id:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/dtmf:
    post:
      tags: [ Calls ]
      x-requirement-id: MEDIA-06
      parameters:
        - $ref: "#/components/parameters/CallId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                leg_id:
                  type: string
                digit:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
...
  /api/v1/calls/{callId}/legs/{legId}/mute:
    post:
      tags: [ MediaSignaling ]
      x-requirement-id: MEDIA-05
      parameters:
        - $ref: "#/components/parameters/CallId"
        - $ref: "#/components/parameters/LegId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                audio:
                  type: boolean
                video:
                  type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/legs/{legId}/offer:
    post:
      tags: [ MediaSignaling ]
      summary: 提交 SDP Offer 或获取 Offer
      x-requirement-id: MEDIA-01,VIDEO-01
      parameters:
        - $ref: "#/components/parameters/CallId"
        - $ref: "#/components/parameters/LegId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SDPPayload"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/SDPPayload"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/legs/{legId}/answer:
    post:
      tags: [ MediaSignaling ]
      parameters:
        - $ref: "#/components/parameters/CallId"
        - $ref: "#/components/parameters/LegId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SDPPayload"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/legs/{legId}/ice:
    post:
      tags: [ MediaSignaling ]
      x-requirement-id: MEDIA-04
      parameters:
        - $ref: "#/components/parameters/CallId"
        - $ref: "#/components/parameters/LegId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ICECandidate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/calls/{callId}/turn-credentials:
    get:
      tags: [ MediaSignaling ]
      x-requirement-id: VIDEO-12,MEDIA-04b
      parameters:
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/TURNCredentials"
                  request_id:
                    type: string
...
  /api/v1/supervisor/agents/{agentId}/force-check-out:
    post:
      tags: [ Supervisor ]
      x-requirement-id: AGENT-05
      x-roles: [ supervisor, admin ]
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
  /api/v1/supervisor/calls/{callId}/listen:
    post:
      tags: [ Supervisor ]
      x-requirement-id: CALL-04
      x-roles: [ supervisor ]
      parameters:
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: 返回 supervisor leg 信令入口信息
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                  - request_id
                properties:
                  code:
                    type: string
                    const: OK
                  message:
                    type: string
                  data:
                    type: "null"
                  request_id:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
...
    CallId:
      name: callId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    LegId:
      name: legId
      in: path
      required: true
      schema:
        type: string
  responses:
...
    BusinessError:
      description: 业务错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required:
        - code
        - message
        - data
        - request_id
        - error
      properties:
        code:
          type: string
        message:
          type: string
        data: {}
        request_id:
          type: string
        error:
          type: string
          description: 错误分类；兼容旧版客户端
...
    Call:
      type: object
      properties:
        id:
          type: string
          format: uuid
        state:
          type: string
        direction:
          type: string
          enum: [ inbound, outbound, internal ]
        session_type:
          type: string
          enum: [ audio, video, mixed ]
        queue_id:
          type: string
        agent_id:
          type: string
        legs:
          type: array
          items:
            $ref: "#/components/schemas/CallLeg"
    CallLeg:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
        agent_id:
          type: string
...
    OutboundCallRequest:
      type: object
      required: [ destination ]
      properties:
        destination:
          type: string
          description: 分机或 E.164
    TransferRequest:
      type: object
      required: [ mode ]
      properties:
        mode:
          type: string
          enum: [ blind, consult ]
        target_agent_id:
          type: string
        target_queue_id:
          type: string
    SDPPayload:
      type: object
      properties:
        sdp:
          type: string
        type:
          type: string
          enum: [ offer, answer ]
    ICECandidate:
      type: object
      properties:
        candidate:
          type: string
        sdp_mid:
          type: string
        sdp_mline_index:
          type: integer
    TURNCredentials:
      type: object
      properties:
        urls:
          type: array
          items:
            type: string
        username:
          type: string
        credential:
          type: string
        ttl:
          type: integer
...